explicitly make moves and captures 16 bits long

+++ include/move.h
@@ -32,7 +32,7 @@ const std::array<std::string, NUM_SQUARES> SquareStr {
 
 // Captures are a bit field composed of two fields
 // [spare : 2][piece : 7][square : 7]
-using Capture = unsigned int;
+using Capture = uint16_t;
 
 constexpr Capture makeCapture(Piece p, Square s) {
   return p << 7 | s;
@@ -52,7 +52,7 @@ inline Square getCapSquare(Capture cptr) {
 // to have its lowest bit treated as a value for
 // positive or negative rotation, else the first 7 bits
 // are to be treated as a board square.
-using Move = unsigned int;
+using Move = uint16_t;
 
 constexpr Move makeMove(Square from, Square to, bool swap=false) {
   return from << 9 | swap << 7 | to;
